Update hr.json
When you read this, you'll see that I make mistakes when writing English, but I understand English text and I know Croatian well. I localize, for example, Notepad++ and Total Commander.

303 - 306
not in the previous translation

030, 032, 033, 034, 149, 150, 246, 247, 297
: null, // missing translation

008, 010, 011, 012, 013, 014, 015, 016, 017, 018, 019, 020, 029, 045, 047, 048, 050, 051, 052, 053, 054, 056, 065, 068, 111, 141, 166, 168, 183, 229, 244
In Croatian localization, it is standard for a human (user) to address a computer (program) with commands, for example "Klikni ovdje".
While software addresses the user respectfully, for example, "Kliknite ovdje", when giving information, queries or warnings. In English, it is the same: "Click here".
I have changed these lines so that it is clear whether it is an instruction, a command by which the user addresses the program, or it is advice, recommendation, notification, warning, when the program addresses the user.

138, 187, 286, 287, 291, 316
"da biste", characteristic for the Serbian language, has been changed to "kako biste", characteristic for the Croatian language.

010, 011, 014, 015 u -> na
053, 087 na -> u
038 iz -> sa
grammatical errors in prepositions (like English "at, in, on...")

255, 256, 262, 268
spelling errors or typos

021, 022, 049, 208, 209, 217
changed according to the terms used in Microsoft software localization, according to "Microsoft Terminology Search"

In the lines 057, 058, 059, 060 "Tag" translated as "Oznaka"
In the lines 055, 056. 063, 064, 222 "Bookmarks" translated as "Oznake"
In the line 060 "Bookmarks" translated as "Značajke"
I translated "Bookmarks" everywhere as "Oznake", as it is in Chrome, for example, while I translated "Tag" everywhere as "Značka"

069, 259
"pogleda čitatelja" -> "prikaz čitača", as in one line above 068, 258

182
space at the end

The rest are not important errors, but it looks better this way...

093, 094, 221
"fokusa" -> "fokusiranja"

115, 121, 128, 130
"Dogodila se pogreška" -> "Došlo je do greške"

019, 027, 028, 077, 116, 117, 125, 132, 134, 192, 196, 201, 203, 229, 249, 271, 290